# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger definition for MedRoom backend.",
        "title": "MedRoom API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Martín Saavedra",
            "url": "https://www.swagger.io/support",
            "email": "martin.saavedra@mail.udp.cl"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/evaluadores": {
            "get": {
                "description": "Lista todos los evaluadores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Lista de evaluadores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SwaggerMessages.ListEvaluadoresSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Genera un nuevo evaluador con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Agrega un nuevo evaluador",
                "parameters": [
                    {
                        "description": "Evaluador a agregar",
                        "name": "input_evaluador",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.AddNewEvaluadorPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.AddNewEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/evaluadores/me": {
            "get": {
                "description": "Obtiene el perfil del evaluador según su token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Obtiene el perfil del evaluador",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetMyEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica el perfil del propio evaluador con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Modifica mi perfil",
                "parameters": [
                    {
                        "description": "Evaluador a modificar",
                        "name": "input_actualiza_evaluador",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutMyEvaluadorPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutMyEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/evaluadores/{uuid_evaluador}": {
            "get": {
                "description": "Obtiene un evaluador según su UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Obtiene un evaluador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del evaluador a buscar",
                        "name": "uuid_evaluador",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.GetOneEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica un evaluador con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Modifica un evaluador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del evaluador a modificar",
                        "name": "uuid_evaluador",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Evaluador a modificar",
                        "name": "input_actualiza_evaluador",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestMessages.PutOneEvaluadorPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.PutOneEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un evaluador con los datos entregados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Evaluadores"
                ],
                "summary": "Elimina un evaluador",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del evaluador a eliminar",
                        "name": "uuid_evaluador",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SwaggerMessages.DeleteEvaluadorSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/ApiHelpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ApiHelpers.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.AddNewEvaluadorPayload": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "hash_contrasena_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutMyEvaluadorPayload": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "hash_contrasena_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "RequestMessages.PutOneEvaluadorPayload": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "hash_contrasena_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.AddNewEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.DeleteEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetMyEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rol_evaluador": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rol_evaluador": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.GetOneRolResponse": {
            "type": "object",
            "properties": {
                "nombre_rol": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.ListEvaluadoresResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rol_evaluador": {
                    "$ref": "#/definitions/ResponseMessages.GetOneRolResponse"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutMyEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "hash_contrasena_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "ResponseMessages.PutOneEvaluadorResponse": {
            "type": "object",
            "properties": {
                "apellidos_evaluador": {
                    "type": "string"
                },
                "cargo_evaluador": {
                    "type": "string"
                },
                "correo_electronico_evaluador": {
                    "type": "string"
                },
                "hash_contrasena_evaluador": {
                    "type": "string"
                },
                "id_rol": {
                    "type": "integer"
                },
                "nombres_evaluador": {
                    "type": "string"
                },
                "recinto_evaluador": {
                    "type": "string"
                },
                "rut_evaluador": {
                    "type": "string"
                },
                "telefono_celular_evaluador": {
                    "type": "string"
                },
                "telefono_fijo_evaluador": {
                    "type": "string"
                }
            }
        },
...
        "SwaggerMessages.AddNewEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.AddNewEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.DeleteEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.DeleteEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetMyEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetMyEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.GetOneEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.GetOneEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.ListEvaluadoresSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.ListEvaluadoresResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutMyEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutMyEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "SwaggerMessages.PutOneEvaluadorSwagger": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ResponseMessages.PutOneEvaluadorResponse"
                },
                "meta": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}